feat: enhance documentation and add new examples for TypeScript utilities
- Added a new script in `package.json` for generating documentation locally.
- Updated `docs.yaml` to include a theme color for better branding.
- Introduced a new `Example.vue` component for showcasing TypeScript utility functions.
- Removed outdated `Example.vue` component to streamline documentation.
- Enhanced multiple API documentation files with improved descriptions and examples.
- Updated the SQLite database to reflect the latest documentation changes.
