+++ docs/.config/docs.yaml
@@ -2,6 +2,7 @@
 
 name: "knitwork-x"
 shortDescription: "Utilities to generate JavaScript code."
+themeColor: "blue"
 description: "Utilities to generate JavaScript code. Forked from knitwork, with comprehensive TypeScript generation tools."
 github: "hairyf/knitwork-x"
 url: "https://knitwork-x.vercel.app"
