+++ docs/2.apis/.navigation.yml
@@ -1,2 +1,2 @@
+title: APIs
 icon: i-lucide-book-open
-name: APIs
