+++ docs/2.apis/1.string.md
@@ -1,12 +1,12 @@
 ---
 title: string
-description: 字符串与字面量生成（genString、escapeString、genSafeVariableName、genRegExp、genTemplateLiteral）
+description: String and literal generation (genString, escapeString, genSafeVariableName, genRegExp, genTemplateLiteral)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成 JavaScript/TypeScript 字符串字面量、转义、安全变量名、正则字面量以及模板字符串的 API。
+This module provides APIs for generating JavaScript/TypeScript string literals, escaping, safe variable names, regex literals, and template strings.
 
-<Example module="string" />
-
+<example module="string"></example>
 <!-- automd:jsdocs src="../../src/string.ts" -->
 
 ## String
