+++ docs/2.apis/10.decorator.md
@@ -1,11 +1,12 @@
 ---
 title: decorator
-description: 装饰器生成（genDecorator）
+description: Decorator generation (genDecorator)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成装饰器语法 `@Decorator` 或 `@Decorator(args)` 的 API。
+This module provides APIs for generating decorator syntax `@Decorator` or `@Decorator(args)`.
 
-<Example module="decorator" />
+<example module="decorator"></example>
 
 <!-- automd:jsdocs src="../../src/decorator.ts" -->
