+++ docs/2.apis/11.module.md
@@ -1,11 +1,12 @@
 ---
 title: module
-description: declare module 与模块增强（genModule、genAugmentation）
+description: declare module and module augmentation (genModule, genAugmentation)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成 TypeScript `declare module` 模块增强（augmentation）的 API。
+This module provides APIs for generating TypeScript `declare module` augmentations.
 
-<Example module="module" />
+<example module="module"></example>
 
 <!-- automd:jsdocs src="../../src/module.ts" -->
