+++ docs/2.apis/12.namespace.md
@@ -1,11 +1,12 @@
 ---
 title: namespace
-description: namespace 与 declare global（genNamespace、genDeclareNamespace）
+description: namespace and declare global (genNamespace, genDeclareNamespace)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成 TypeScript `namespace` 与 `declare global` 等声明块的 API。
+This module provides APIs for generating TypeScript `namespace` and `declare global` blocks.
 
-<Example module="namespace" />
+<example module="namespace"></example>
 
 <!-- automd:jsdocs src="../../src/namespace.ts" -->
