+++ docs/2.apis/13.condition.md
@@ -1,11 +1,12 @@
 ---
 title: condition
-description: 条件与分支生成（genIf、genElseIf、genElse、genTernary、genPrefixedBlock）
+description: Condition and branch generation (genIf, genElseIf, genElse, genTernary, genPrefixedBlock)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成 `if` / `else if` / `else` 以及三元表达式、通用前缀块（如 `while`）的 API。
+This module provides APIs for generating `if` / `else if` / `else`, ternary expressions, and generic prefixed blocks (e.g. `while`).
 
-<Example module="condition" />
+<example module="condition"></example>
 
 <!-- automd:jsdocs src="../../src/condition.ts" -->
