+++ docs/2.apis/14.try.md
@@ -1,11 +1,12 @@
 ---
 title: try
-description: try/catch/finally 生成（genTry、genCatch、genFinally）
+description: try/catch/finally generation (genTry, genCatch, genFinally)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成 `try`、`catch`、`finally` 语句块的 API，支持可选的 catch 绑定。
+This module provides APIs for generating `try`, `catch`, and `finally` blocks, with optional catch binding support.
 
-<Example module="try" />
+<example module="try"></example>
 
 <!-- automd:jsdocs src="../../src/try.ts" -->
