+++ docs/2.apis/15.loop.md
@@ -1,11 +1,12 @@
 ---
 title: loop
-description: 循环生成（genFor、genForOf、genForIn、genWhile、genDoWhile）
+description: Loop generation (genFor, genForOf, genForIn, genWhile, genDoWhile)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成 `for`、`for...of`、`for...in`、`while`、`do...while` 循环的 API。
+This module provides APIs for generating `for`, `for...of`, `for...in`, `while`, and `do...while` loops.
 
-<Example module="loop" />
+<example module="loop"></example>
 
 <!-- automd:jsdocs src="../../src/loop.ts" -->
