+++ docs/2.apis/16.switch.md
@@ -1,11 +1,12 @@
 ---
 title: switch
-description: switch/case 生成（genSwitch、genCase、genDefault）
+description: switch/case generation (genSwitch, genCase, genDefault)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成 `switch` 语句及 `case` / `default` 分支的 API。
+This module provides APIs for generating `switch` statements and `case` / `default` branches.
 
-<Example module="switch" />
+<example module="switch"></example>
 
 <!-- automd:jsdocs src="../../src/switch.ts" -->
