+++ docs/2.apis/17.statement.md
@@ -1,11 +1,12 @@
 ---
 title: statement
-description: 语句生成（genReturn、genThrow）
+description: Statement generation (genReturn, genThrow)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成 `return` 与 `throw` 语句的 API。
+This module provides APIs for generating `return` and `throw` statements.
 
-<Example module="statement" />
+<example module="statement"></example>
 
 <!-- automd:jsdocs src="../../src/statement.ts" -->
