+++ docs/2.apis/18.object.md
@@ -1,11 +1,12 @@
 ---
 title: object
-description: 对象与集合字面量生成（genObjectLiteral、genObjectFromRaw、genArrayFromRaw、genMapFromRaw、genSetFromRaw）
+description: Object and collection literal generation (genObjectLiteral, genObjectFromRaw, genArrayFromRaw, genMapFromRaw, genSetFromRaw)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成对象字面量、数组、Map、Set 以及键值对序列化的 API，支持原始值与带引号的值。
+This module provides APIs for generating object literals, arrays, Map, Set, and key-value serialization, with support for raw and quoted values.
 
-<Example module="object" />
+<example module="object"></example>
 
 <!-- automd:jsdocs src="../../src/object.ts" -->
