+++ docs/2.apis/19.utils.md
@@ -1,11 +1,12 @@
 ---
 title: utils
-description: 工具函数（genJSDocComment、wrapInDelimiters、genObjectKey、genComment）
+description: Utility functions (genJSDocComment, wrapInDelimiters, genObjectKey, genComment)
+icon: i-lucide-square-function
 ---
 
-本模块提供 JSDoc 注释、分隔符包裹、对象键安全生成与单行/块注释等通用工具 API。
+This module provides utility APIs for JSDoc comments, delimiter wrapping, safe object key generation, and single-line/block comments.
 
-<Example module="utils" />
+<example module="utils"></example>
 
 <!-- automd:jsdocs src="../../src/utils.ts" -->
