+++ docs/2.apis/2.variable.md
@@ -1,11 +1,12 @@
 ---
 title: variable
-description: 变量声明生成（genVariable）
+description: Variable declaration generation (genVariable)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成变量声明（`const` / `let`）的 API，支持导出与 kind 选项。
+This module provides APIs for generating variable declarations (`const` / `let`) with export and kind options.
 
-<Example module="variable" />
+<example module="variable"></example>
 
 <!-- automd:jsdocs src="../../src/variable.ts" -->
