+++ docs/2.apis/3.esm.md
@@ -1,11 +1,12 @@
 ---
 title: ESM
-description: ESM 模块语法生成（import / export / 动态 import）
+description: ESM module syntax generation (import / export / dynamic import)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成 ESM 导入、导出、默认导出、动态 import 及 `typeof import()` 的 API，支持 import attributes。
+This module provides APIs for generating ESM imports, exports, default exports, dynamic import, and `typeof import()`, with support for import attributes.
 
-<Example module="esm" />
+<example module="esm"></example>
 
 <!-- automd:jsdocs src="../../src/esm.ts" -->
