+++ docs/2.apis/4.class.md
@@ -1,11 +1,12 @@
 ---
 title: class
-description: 类与类成员生成（genClass、genConstructor、genClassProperty、genClassMethod、get/set）
+description: Class and class member generation (genClass, genConstructor, genClassProperty, genClassMethod, get/set)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成 TypeScript/JavaScript 类、构造函数、类属性、类方法以及 getter/setter 的 API。
+This module provides APIs for generating TypeScript/JavaScript classes, constructors, class properties, class methods, and getters/setters.
 
-<Example module="class" />
+<example module="class"></example>
 
 <!-- automd:jsdocs src="../../src/class.ts" -->
