+++ docs/2.apis/5.interface.md
@@ -1,11 +1,12 @@
 ---
 title: interface
-description: 接口与类型签名生成（genInterface、genIndexSignature、genCallSignature、genConstructSignature）
+description: Interface and type signature generation (genInterface, genIndexSignature, genCallSignature, genConstructSignature)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成 TypeScript 接口、索引签名、调用签名与构造签名的 API。
+This module provides APIs for generating TypeScript interfaces, index signatures, call signatures, and construct signatures.
 
-<Example module="interface" />
+<example module="interface"></example>
 
 <!-- automd:jsdocs src="../../src/interface.ts" -->
