+++ docs/2.apis/6.enum.md
@@ -1,11 +1,12 @@
 ---
 title: enum
-description: 枚举生成（genEnum、genConstEnum）
+description: Enum generation (genEnum, genConstEnum)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成 TypeScript 枚举与 const 枚举的 API，支持数字、字符串成员及自动递增。
+This module provides APIs for generating TypeScript enums and const enums, with support for numeric, string members, and auto-increment.
 
-<Example module="enum" />
+<example module="enum"></example>
 
 <!-- automd:jsdocs src="../../src/enum.ts" -->
