+++ docs/2.apis/7.function.md
@@ -1,11 +1,12 @@
 ---
 title: function
-description: 函数与块生成（genFunction、genArrowFunction、genMethod、genBlock、genParam）
+description: Function and block generation (genFunction, genArrowFunction, genMethod, genBlock, genParam)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成函数声明、箭头函数、对象方法、语句块与参数列表的 API，支持泛型、async、generator 与 JSDoc。
+This module provides APIs for generating function declarations, arrow functions, object methods, statement blocks, and parameter lists, with support for generics, async, generator, and JSDoc.
 
-<Example module="function" />
+<example module="function"></example>
 
 <!-- automd:jsdocs src="../../src/function.ts" -->
