+++ docs/2.apis/8.type-alias.md
@@ -1,11 +1,12 @@
 ---
 title: type-alias
-description: 类型别名与类型表达式（genTypeAlias、genTypeObject、genUnion、genIntersection、genMappedType 等）
+description: Type alias and type expression generation (genTypeAlias, genTypeObject, genUnion, genIntersection, genMappedType, etc.)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成类型别名、对象类型、联合/交叉类型、映射类型、模板字面量类型以及 keyof/typeof/断言/satisfies 的 API。
+This module provides APIs for generating type aliases, object types, union/intersection types, mapped types, template literal types, and keyof/typeof/assertion/satisfies.
 
-<Example module="type-alias" />
+<example module="type-alias"></example>
 
 <!-- automd:jsdocs src="../../src/type-alias.ts" -->
