+++ docs/2.apis/9.conditional.md
@@ -1,11 +1,12 @@
 ---
 title: conditional
-description: 条件类型生成（genConditionalType）
+description: Conditional type generation (genConditionalType)
+icon: i-lucide-square-function
 ---
 
-本模块提供生成 TypeScript 条件类型 `T extends U ? X : Y` 的 API。
+This module provides APIs for generating TypeScript conditional types `T extends U ? X : Y`.
 
-<Example module="conditional" />
+<example module="conditional"></example>
 
 <!-- automd:jsdocs src="../../src/conditional.ts" -->
