+++ package.json
@@ -22,6 +22,7 @@
   "scripts": {
     "build": "unbuild",
     "dev": "vitest dev --coverage",
+    "docs": "cd docs && pnpm dev",
     "lint": "eslint --cache . && prettier -c src test",
     "lint:fix": "automd && eslint --cache . --fix && prettier -c src test -w",
     "prepack": "pnpm run build",
